test: move general 'rpc' pkg test logic to a new  'main_test.go' file
